Implemented better response for functions

+++ src/main/kotlin/com/acme/tour/controller/PromocaoController.kt
@@ -1,9 +1,13 @@
 package com.acme.tour.controller
 
 import com.acme.tour.model.Promocao
+import com.acme.tour.model.RespostaJSON
 import com.acme.tour.service.PromocaoService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
+import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 
 @RestController
@@ -14,25 +18,49 @@ class PromocaoController {
     lateinit var promocaoService: PromocaoService
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable id: Long) = promocaoService.getById(id)
+    fun getById(@PathVariable id: Long) : ResponseEntity<Promocao?> {
+        var promocao = promocaoService.getById(id)
+
+        var status = if(promocao == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+        return ResponseEntity(promocao, status)
+    }
 
     @PostMapping()
-    fun create(@RequestBody promocao: Promocao){
+    fun create(@RequestBody promocao: Promocao): ResponseEntity<RespostaJSON>{
         promocaoService.create(promocao)
+        val respostaJSON = RespostaJSON("OK", Date())
+        return ResponseEntity(respostaJSON, HttpStatus.CREATED)
     }
 
     @DeleteMapping("/{id}")
-    fun delete(@PathVariable id:Long){
-        promocaoService.delete(id)
+    fun delete(@PathVariable id:Long): ResponseEntity<Unit>{
+        var status = HttpStatus.NOT_FOUND
+        if(promocaoService.getById(id) != null){
+            status = HttpStatus.ACCEPTED
+            promocaoService.delete(id)
+        }
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping("/{id}")
-    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao){
-        promocaoService.update(id, promocao)
+    fun update(@PathVariable id: Long, @RequestBody promocao: Promocao): ResponseEntity<Unit>{
+
+        var status = HttpStatus.NOT_FOUND
+        if(promocaoService.getById(id) != null){
+            promocaoService.update(id, promocao)
+            status = HttpStatus.ACCEPTED
+        }
+        return ResponseEntity(Unit, status)
     }
 
     @GetMapping()
-    fun getAll(@RequestParam(required = false, defaultValue = "") localFilter: String) =
-        promocaoService.searchByLocal(localFilter)
+    fun getAll(@RequestParam(required = false, defaultValue = "") localFilter: String): ResponseEntity<List<Promocao>> {
+        var status = HttpStatus.OK
+        var listaPromocoes = promocaoService.searchByLocal(localFilter)
+        if(listaPromocoes.size == 0){
+            status = HttpStatus.NOT_FOUND
+        }
+        return ResponseEntity(listaPromocoes, status)
+    }
 
 }
